+++ src/main/java/cz/preclik/shop/preclikshop/jpa/PriceRepository.java
@@ -4,6 +4,6 @@ import cz.preclik.shop.preclikshop.domain.Price;
 import cz.preclik.shop.preclikshop.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PriceRepository extends JpaRepository<Price, Integer> {
+public interface PriceRepository extends JpaRepository<Price, Long> {
     Price findFirstByProductEqualsOrderByValidFromDesc(final Product product);
 }
