+++ src/main/java/cz/preclik/shop/preclikshop/jpa/ProductRepository.java
@@ -6,12 +6,12 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
     @Modifying
     @Query("update Product p set p.available = :available where p.id = :id")
     Integer setAvailable(@Param("available") final Boolean available, @Param("id") Integer id);
 
     @Modifying
     @Query("update Product p set p.quantity = p.quantity + :quantity WHERE p.id = :id")
-    Integer increaseQuantity(@Param("quantity") final Integer quantity, @Param("id") Integer id);
+    Integer increaseQuantity(@Param("quantity") final Integer quantity, @Param("id") Long id);
 }
