+++ src/main/java/cz/preclik/shop/preclikshop/service/NegativeQuantityOfProductException.java
@@ -1,7 +1,7 @@
 package cz.preclik.shop.preclikshop.service;
 
 public class NegativeQuantityOfProductException extends Throwable {
-    public NegativeQuantityOfProductException(final Integer id){
+    public NegativeQuantityOfProductException(final Long id){
         super("Negative quantity of " + id +  " product");
     }
 }
